MAINT: renaming string.h to xstring.h
* See issue #91
* This fix was identified by @torognes
