+++ src/xstring.h
